Move ProjectTimeForm to route file
