+++ app/components/ProjectTimes.tsx
@@ -1,8 +1,7 @@
-import { convertFloatTimeToHHMM } from "~/utils/dateTimeUtils";
-import { ProjectTimeForm } from "./ProjectTimeForm";
 import { TrackyTask } from "~/apis/tasks/TrackyTask";
-import { CalculationPosition } from "~/apis/troi/troi.types";
-import { TroiProjectTime } from "~/apis/troi/troi.types";
+import { CalculationPosition, TroiProjectTime } from "~/apis/troi/troi.types";
+import { convertFloatTimeToHHMM } from "~/utils/dateTimeUtils";
+import { ProjectTimeForm } from "../routes/project_time.($id)";
 
 interface Props {
   selectedDate: Date;
